requirements.txt
Flask==2.2
Flask-SQLAlchemy==3.0.2
# psycopg2-binary==2.9.1
redis==3.5.3
Flask-Migrate==3.1.0
Flask-WTF==0.15.1
Flask-Validator==1.4
loguru==0.5.3
Flask-Mail==0.9.1
Flask-Admin==1.5.8
Flask-Login==0.6.2
Flask-Security==3.0.0
Flask-Bootstrap==3.3.7.0
pydantic==1.8.2
Flask-Babel==2.0.0
humanize==3.11.0
bcrypt==3.2.0
itsdangerous==2.0.1
gunicorn==20.1.0
pillow==8.3.2
selenium==3.14
werkzeug==2.2.0
jinja2==3.0.3
pytelegrambotapi==4.12.0
dpath==2.1.6
sqlalchemy==1.4.49